Refactor to fix build
